Aggiunti repository, handler e controller per carrello e implementata logica di interazione dell'acquirente con esso

+++ FilieraFanesiCardinali/src/main/java/unicam/filierafanesicardinali/repository/CuratoreRepository.java
@@ -3,5 +3,5 @@ package unicam.filierafanesicardinali.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import unicam.filierafanesicardinali.model.amministrazione.Curatore;
 
-public interface CuratoreRepository extends JpaRepository<Curatore, Integer> {
+public interface CuratoreRepository extends JpaRepository<Curatore, Long> {
 }
