+++ FilieraFanesiCardinali/src/main/java/unicam/filierafanesicardinali/repository/EventoRepository.java
@@ -3,5 +3,5 @@ package unicam.filierafanesicardinali.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import unicam.filierafanesicardinali.model.eventi.Evento;
 
-public interface EventoRepository extends JpaRepository<Evento, Integer> {
+public interface EventoRepository extends JpaRepository<Evento, Long> {
 }
